@@ -1,25 +1,19 @@
 {
   "name": "personal-portfolio",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.19",
-    "@fortawesome/free-brands-svg-icons": "^5.9.0",
-    "@fortawesome/free-regular-svg-icons": "^5.9.0",
-    "@fortawesome/free-solid-svg-icons": "^5.9.0",
-    "@fortawesome/react-fontawesome": "^0.1.4",
-    "@material-ui/core": "^4.0.2",
-    "@material-ui/icons": "^4.0.1",
-    "@material-ui/lab": "^4.0.0-alpha.14",
-    "core-js": "^3.19.3",
-    "querystringify": "^2.1.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-pdf": "^4.0.5",
-    "react-player": "^1.11.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^4.0.3",
-    "yarn": "^1.22.17"
+    "@emotion/react": "^11.8.2",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.5.1",
+    "firebase": "^9.6.9",
+    "npm": "^8.5.5",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
+    "react-router-dom": "^6.2.2",
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
