@@ -1,72 +1,54 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
+import * as React from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
+import Menu from "@mui/material/Menu";
+import MenuIcon from "@mui/icons-material/Menu";
+import Container from "@mui/material/Container";
+import Avatar from "@mui/material/Avatar";
+import Button from "@mui/material/Button";
+import Tooltip from "@mui/material/Tooltip";
+import MenuItem from "@mui/material/MenuItem";
 
-import Drawer from '../modules/Drawer'
-import { Link } from 'react-router-dom'
-import { Hidden } from '@material-ui/core';
+import Drawer from "../modules/Drawer"
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-    textDecoration: 'none',
-  },
-}));
 
-const LandingPage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/" {...props} />
-));
-const AboutPage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/about" {...props} />
-));
-const ContactPage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/contact" {...props} />
-));
-const BlogPage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/blog" {...props} />
-));
-const ResumePage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/resume" {...props} />
-));
-const ProjectsPage = React.forwardRef((props, ref) => (
-  <Link innerRef={ref} to="/projects" {...props} />
-));
+const NavBar = () => {
+  const pages = ["Products", "Pricing", "Blog"];
 
-function NavBar() {
-  const classes = useStyles();
   return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          {/* <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="Menu"> */}
-            <Drawer />
-            {/* <MenuIcon /> */}
-          {/* </IconButton> */}
-          <Typography component={LandingPage} variant="h6" color="inherit" className={classes.title}>
-            {/* <Button className={classes.title} component={LandingPage} color="inherit">Home</Button> */}
-            Long Pham
-          </Typography>
-          <Hidden smDown>
-            <Button component={AboutPage} color="inherit">About Me</Button>         
-            <Button component={ProjectsPage} color="inherit">Projects</Button>
-            <Button component={BlogPage} color="inherit">Blog</Button>
-            <Button component={ResumePage} color="inherit">Resume/CV</Button>
-            {/* <Button disabled component={ContactPage} color="inherit">Contact</Button> */}
-            <Button component={ContactPage} color="inherit">Contact</Button>
-          </Hidden>
-        </Toolbar>
-      </AppBar>
-    </div>
+    <AppBar position="static">
+      <Toolbar>
+        <Drawer />
+        <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+          News
+        </Typography>
+        <Button color="inherit">Login</Button>
+      </Toolbar>
+    </AppBar>
+
+    // <>
+    //   <AppBar position="static">
+    //     <Toolbar>
+    //       <Drawer />
+    //       <Typography variant="h6" color="inherit">
+    //         {/* <Button className={classes.title} component={LandingPage} color="inherit">Home</Button> */}
+    //         Long Pham
+    //       </Typography>
+    //       <Hidden smDown>
+    //         <Button component={AboutPage} color="inherit">About Me</Button>
+    //         <Button component={ProjectsPage} color="inherit">Projects</Button>
+    //         <Button component={BlogPage} color="inherit">Blog</Button>
+    //         <Button component={ResumePage} color="inherit">Resume/CV</Button>
+    //         {/* <Button disabled component={ContactPage} color="inherit">Contact</Button> */}
+    //         <Button component={ContactPage} color="inherit">Contact</Button>
+    //       </Hidden>
+    //     </Toolbar>
+    //   </AppBar>
+    // </>
   );
-}
+};
 
 export default NavBar;
